Collapsed AutoDateHistogramAggregation and AutoHistogramAggregationBase into one class, cleaned up available options.
