+++ src/aggregations/bucket-aggregations/index.js
@@ -2,7 +2,6 @@
 
 exports.BucketAggregationBase = require('./bucket-aggregation-base');
 exports.HistogramAggregationBase = require('./histogram-aggregation-base');
-exports.AutoHistogramAggregationBase = require('./auto-histogram-aggregation-base');
 exports.RangeAggregationBase = require('./range-aggregation-base');
 exports.TermsAggregationBase = require('./terms-aggregation-base');
 exports.SignificantAggregationBase = require('./significant-aggregation-base');
