+++ src/index.d.ts
@@ -4919,12 +4919,10 @@ declare namespace esb {
     }
 
     /**
-     * The `AutoHistogramAggregationBase` provides support for common options used across
-     * various auto histogram `Aggregation` implementations like Auto Date Histogram aggregation.
-     *
-     * **NOTE:** Instantiating this directly should not be required.
-     * However, if you wish to add a custom implementation for whatever reason,
-     * this class could be extended.
+     * The `AutoDateHistogramAggregation` is similar to the Date histogram aggregation except 
+     * instead of providing an interval to use as the width of each bucket, a target number 
+     * of buckets is provided indicating the number of buckets needed and the interval of the 
+     * buckets is automatically chosen to best achieve that target. 
      *
      * @param {string} name The name which will be used to refer to this aggregation.
      * @param {string} aggType Type of aggregation
@@ -4932,16 +4930,15 @@ declare namespace esb {
      * @param {number=} buckets Bucket count to generate histogram over.
      * @extends BucketAggregationBase
      */
-    export class AutoHistogramAggregationBase extends BucketAggregationBase {
+    export class AutoDateHistogramAggregation extends BucketAggregationBase {
         constructor(
             name: string,
-            aggType: string,
             field?: string,
             buckets?: number
         );
 
         /**
-         * Sets the histogram interval. Buckets are generated based on this interval value.
+         * Sets the bucket count. Buckets are generated based on this interval value.
          *
          * @param {number} buckets Bucket count to generate histogram over.
          */
@@ -4957,51 +4954,6 @@ declare namespace esb {
         format(fmt: string): this;
 
         /**
-         * The offset parameter is used to change the start value of each bucket
-         * by the specified positive (+) or negative offset (-).
-         * Negative offset is not applicable on HistogramAggregation.
-         * In case of DateHistogramAggregation, duration can be
-         * a value such as 1h for an hour, or 1d for a day.
-         *
-         * @param {string} offset Time or bucket key offset for bucketing.
-         */
-        offset(offset: string): this;
-
-        /**
-         * Sets the ordering for buckets
-         *
-         * @param {string} key
-         * @param {string} direction `asc` or `desc`
-         */
-        order(key: string, direction?: 'asc' | 'desc'): this;
-
-        /**
-         * Sets the minimum number of matching documents in range to return the bucket.
-         *
-         * @param {number} minDocCnt Integer value for minimum number of documents
-         * required to return bucket in response
-         */
-        minDocCount(minDocCnt: number): this;
-
-        /**
-         * Set's the range/bounds for the histogram aggregation.
-         * Useful when you want to include buckets that might be
-         * outside the bounds of indexed documents.
-         *
-         * @param {number|string} min Start bound / minimum bound value
-         * For histogram aggregation, Integer value can be used.
-         * For Date histogram, date expression can be used.
-         * Available expressions for interval:
-         * year, quarter, month, week, day, hour, minute, second
-         * @param {number|string} max End bound / maximum bound value
-         * For histogram aggregation, Integer value can be used.
-         * For Date histogram, date expression can be used.
-         * Available expressions for interval:
-         * year, quarter, month, week, day, hour, minute, second
-         */
-        extendedBounds(min: number | string, max: number | string): this;
-
-        /**
          * Sets the missing parameter which defines how documents
          * that are missing a value should be treated.
          *
@@ -5010,25 +4962,11 @@ declare namespace esb {
         missing(value: string): this;
 
         /**
-         * Enable the response to be returned as a keyed object where the key is the
-         * bucket interval.
+         * Sets the minimum rounding interval that should be used.
          *
-         * @param {boolean} keyed To enable keyed response or not.
+         * @param {string} value
          */
-        keyed(keyed: boolean): this;
-    }
-
-    /**
-     * A multi-bucket aggregation similar to the histogram except it can only be applied on date values.
-     * The interval can be specified by date/time expressions.
-     *
-     * @param {string} name The name which will be used to refer to this aggregation.
-     * @param {string=} field The field to aggregate on
-     * @param {number=} buckets Bucket count to generate histogram over.
-     * @extends AutoHistogramAggregationBase
-     */
-    export class AutoDateHistogramAggregation extends AutoHistogramAggregationBase {
-        constructor(name: string, field?: string, buckets?: number);
+        minimumInterval(value: string): this;
 
         /**
          * Date-times are stored in Elasticsearch in UTC.
